add show table source

+++ src/model/table/tableGroup.ts
@@ -52,7 +52,7 @@ export class TableGroup implements INode, IConnection {
     createTemplate() {
         ConnectionManager.getConnection(this, true)
         QueryUnit.createSQLTextDocument(`CREATE TABLE tableName(  
-  id long NOT NULL primary key AUTO_INCREMENT,
+  id int NOT NULL primary key AUTO_INCREMENT,
   column varchar(length)
 );`)
     }
